+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/factoryIrUtil.kt
@@ -15,6 +15,7 @@ import dev.zacsweers.metro.compiler.ir.setDispatchReceiver
 import dev.zacsweers.metro.compiler.ir.stubExpression
 import dev.zacsweers.metro.compiler.ir.thisReceiverOrFail
 import dev.zacsweers.metro.compiler.metroAnnotations
+import org.jetbrains.kotlin.backend.jvm.codegen.AnnotationCodegen.Companion.annotationClass
 import org.jetbrains.kotlin.ir.builders.IrBuilderWithScope
 import org.jetbrains.kotlin.ir.builders.declarations.addFunction
 import org.jetbrains.kotlin.ir.builders.irExprBody
@@ -28,6 +29,7 @@ import org.jetbrains.kotlin.ir.declarations.IrValueParameter
 import org.jetbrains.kotlin.ir.expressions.IrConstructorCall
 import org.jetbrains.kotlin.ir.expressions.IrExpression
 import org.jetbrains.kotlin.ir.symbols.IrConstructorSymbol
+import org.jetbrains.kotlin.ir.util.classId
 import org.jetbrains.kotlin.ir.util.copyAnnotationsFrom
 import org.jetbrains.kotlin.ir.util.copyParametersFrom
 import org.jetbrains.kotlin.ir.util.copyTo
@@ -128,36 +130,47 @@ internal fun generateStaticNewInstanceFunction(
 
 /**
  * Generates a metadata-visible function in the factory class that matches the signature of the
- * target constructor. This function is used in downstream compilations to read the constructor's
- * signature.
+ * target function. This function is used in downstream compilations to read the function's
+ * signature and also dirty IC.
  */
-internal fun generateMetadataVisibleConstructorFunction(
+internal fun generateMetadataVisibleMirrorFunction(
   context: IrMetroContext,
   factoryClass: IrClass,
-  targetConstructor: IrConstructor,
+  target: IrFunction,
 ): IrSimpleFunction {
   val function =
     factoryClass
       .addFunction {
-        name = Symbols.Names.constructorFunction
-        returnType = targetConstructor.returnType
+        name = Symbols.Names.mirrorFunction
+        returnType = target.returnType
       }
       .apply {
-        val sourceClass = factoryClass.parentAsClass
-        val scopeAndQualifierAnnotations = buildList {
-          val classMetroAnnotations = sourceClass.metroAnnotations(context.symbols.classIds)
-          classMetroAnnotations.scope?.ir?.let(::add)
-          classMetroAnnotations.qualifier?.ir?.let(::add)
-        }
-        if (scopeAndQualifierAnnotations.isNotEmpty()) {
-          val container =
-            object : IrAnnotationContainer {
-              override val annotations: List<IrConstructorCall> = scopeAndQualifierAnnotations
+        if (target is IrConstructor) {
+          val sourceClass = factoryClass.parentAsClass
+          val scopeAndQualifierAnnotations = buildList {
+            val classMetroAnnotations = sourceClass.metroAnnotations(context.symbols.classIds)
+            classMetroAnnotations.scope?.ir?.let(::add)
+            classMetroAnnotations.qualifier?.ir?.let(::add)
+          }
+          if (scopeAndQualifierAnnotations.isNotEmpty()) {
+            val container =
+              object : IrAnnotationContainer {
+                override val annotations: List<IrConstructorCall> = scopeAndQualifierAnnotations
+              }
+            copyAnnotationsFrom(container)
+          }
+          copyTypeParametersFrom(sourceClass)
+        } else {
+          // If it's a regular (provides) function, just always copy its annotations
+          // Exclude @Provides to avoid reentrant factory gen
+          // TODO maybe make this more precise in what it copies?
+          copyAnnotationsFrom(target)
+          annotations =
+            annotations.filterNot {
+              it.annotationClass.classId in context.symbols.classIds.providesAnnotations
             }
-          copyAnnotationsFrom(container)
         }
-        copyTypeParametersFrom(sourceClass)
-        copyParametersFrom(targetConstructor)
+        copyParametersFrom(target)
         setDispatchReceiver(factoryClass.thisReceiverOrFail.copyTo(this))
 
         regularParameters.forEach {
@@ -170,7 +183,7 @@ internal fun generateMetadataVisibleConstructorFunction(
               }
           }
         }
-        // The function's signature already matches the target constructor's signature, all we need
+        // The function's signature already matches the target function's signature, all we need
         // this for
         body =
           context.pluginContext.createIrBuilder(symbol).run {